requirements.txt
ai2thor==3.1.0
boto3==1.15.2
fuzzywuzzy==0.18.0
networkx==2.5
pydub==0.24.1
python-Levenshtein
tqdm
pillow
pydantic==1.9.0
numpy
pandas==1.4.2
opencv-python
vocab
revtok
Pillow
sacred
etaprogress
scikit-video
lmdb
gtimer
filelock
termcolor
torch==1.11.0
torchvision==0.12.0
tensorboardX==2.5
Flask
flask_restful
future
pandoc
six
typing
ConfigArgParse
nltk
gensim
jupyter
pytorch-lightning
hydra-core